fix/logrocket-monitoring-FRT-400
